Add files via upload
During the previous semester, I collaborated with a group of five to develop a vintage online shop as part of our Web Programming course (PRG4201) in Inti International University. Using PHP, HTML, CSS, and JavaScript, we worked together to create a functional and visually appealing e-commerce platform. This project was an excellent opportunity to learn the practical aspects of building a commercial website, from frontend design to backend functionality, while fostering teamwork and problem-solving skills with my colleagues.